moved qasm compiler to waveform control CC
