+++ pycqed/tests/test_qasm_compiler_XFU.py
@@ -7,8 +7,9 @@ import numpy as np
 import pycqed as pq
 from io import StringIO
 from pycqed.utilities import general as gen
-from pycqed.instrument_drivers.physical_instruments._controlbox \
-    import qasm_compiler as qcx
+# from pycqed.instrument_drivers.physical_instruments._controlbox \
+#     import qasm_compiler as qcx
+from pycqed.measurement.waveform_control_CC import qasm_compiler as qcx
 from pycqed.instrument_drivers.physical_instruments._controlbox.Assembler \
     import Assembler
 from os.path import join
